requirements.txt
fastapi
pydantic
uvicorn
psycopg2
python-dotenv
requests
yfinance
numpy
pandas

